docs roadmap: add to list of all docs
This avoids an MkDocs warning.

I put it into Design Docs, since people who follow the link to the
roadmap might be interested in those.

It's not a perfect place for it, but it seems better than "Technical
Details" or "Contributing" (which might make people think that they have
to read or follow the roadmap before contributing).