Unset other `GIT_DIR` related env vars in `git config -l`
This has the `git config -l ...` invocation `gix-path` uses to find
a configuration file to treat as associated with the `git`
installation remove more environment variables related to `GIT_DIR`
besides `GIT_COMMON_DIR`.

The new removals probably don't make a difference, because they are
only used if `GIT_DIR` is either unset or set to a location that
actually has a Git repository (that `git` is willing to use).

Unsetting these others is thus more to make clear that they have no
effect than to prevent them from having an effect, though various
versions and downstream releases of `git` exist, so it's narrowly
possible that this is also avoiding a problem somewhere, I guess.

This change is motivated by my realization in 9d4dd12 (#1581) that
it could be slightly beneficial. Other differences between the
variables unset here and in `gix-testtools` remain due to the
differences described in that commit message.