Check for uncommitted unignored generated archives on CI
This adds a step to `test-fast` to run `git diff` and fail if there
are any changes to any tracked files, and sets the newly recognized
`GIX_TEST_CREATE_ARCHIVES_EVEN_ON_CI` environment variable for the
`nextest` step in which fixture scripts are run.

Any changes would indicate a problem that should be addressed, but
the main goal here is to catch when a regenerated archive has not
been committed.

Because adding archives to `.gitignore` would also cause this to
pass but is far less often the appropriate solution, this change
includes a comment about what the best solution *usually* is.

This is added to the `test-fast` jobs but not to `test` job because
the `test` job uses `GIT_TEST_IGNORE_ARCHIVES` to prevent already
existing archives from being used, which will nearly always result
in at least slightly different archives being generated. So in the
`test` job this would, in practice, always give a false-positive
failure.

Because archive generation is (at least currently) suppressed on
Windows, this step should rarely if ever fail on Windows even when
the problem it is looking for is present. But if it does fail on
Windows, then there is a problem that ought to be investigated,
and if the problem it is looking for is present, then it will
still fail on the other systems, unless the problem is specific to
Windows.

Running this step on both Ubuntu and macOS (and Windows), rather
than just Ubuntu (and Windows), has the benefit of catching some
kinds of skew. It is also somewhat simpler than setting it only
for some operating systems.