docs config diff-editor: clean up, change KDiff3 example to Meld
I don't think recommending KDIff3 for diff editing is good advice in
most cases (as opposed to merge conflict resolution, for which KDiff3 is
very good). OTOH, trying Meld for diff editing is a good idea for many
people.

Historically, this part of the docs uses KDiff3 because it was written
when `meld` was the default and `:builtin` did not exist.