Set +x in index in other gix-status fixtures
For these the current situation is simpler because the tests
currently using them do not contain assertions whose correctness
varies based on whether the affected files modes' have +x or -x.

But it is the intent of the fixtures to record +x modes for some
particular files in the test repository, so this modifies the
scripts to do so, even on Windows where `chmod +x` does not take
effect.

This also allows the tests to include a broader range of metadata
and, specifically, may verify in some comparisons that having a
mode of 0o755 instead of 0o644 does not cause a problem, even
though not explicitly asserted in detail.