When a doctest fails, run further doctests
Instead of stopping running doctests even in other packages.

This change only applies to the `test-fast` job on CI. It does not
affect any default configuration for doctests, nor the `justfile`.

The reason for making this change before fixing the remaining
doctest failure (which is in `gix-pack/src/index/mod.rs`) is to
confirm that it takes effect on CI.